@@ -7,63 +7,40 @@ package com.jiusiz.infiray;
 public class ByteUtil {
         /**
          * 转换short为byte
-         *
-         * @param b
-         * @param s
-         *            需要转换的short
-         * @param index
          */
         public static void putShort(byte[] b, short s, int index) {
             b[index + 1] = (byte) (s >> 8);
-            b[index + 0] = (byte) (s >> 0);
+            b[index] = (byte) (s);
         }
 
         /**
          * 通过byte数组取到short
-         *
-         * @param b
-         * @param index
-         *            第几位开始取
-         * @return
          */
         public static short getShort(byte[] b, int index) {
-            return (short) (((b[index + 1] << 8) | b[index + 0] & 0xff));
+            return (short) (((b[index + 1] << 8) | b[index] & 0xff));
         }
 
         /**
          * 转换int为byte数组
-         *
-         * @param bb
-         * @param x
-         * @param index
          */
         public static void putInt(byte[] bb, int x, int index) {
             bb[index + 3] = (byte) (x >> 24);
             bb[index + 2] = (byte) (x >> 16);
             bb[index + 1] = (byte) (x >> 8);
-            bb[index + 0] = (byte) (x >> 0);
+            bb[index] = (byte) (x);
         }
 
         /**
          * 通过byte数组取到int
-         *
-         * @param bb
-         * @param index
-         *            第几位开始
-         * @return
          */
         public static int getInt(byte[] bb, int index) {
             return (int) ((((bb[index + 3] & 0xff) << 24)
                     | ((bb[index + 2] & 0xff) << 16)
-                    | ((bb[index + 1] & 0xff) << 8) | ((bb[index + 0] & 0xff) << 0)));
+                    | ((bb[index + 1] & 0xff) << 8) | ((bb[index] & 0xff))));
         }
 
         /**
          * 转换long型为byte数组
-         *
-         * @param bb
-         * @param x
-         * @param index
          */
         public static void putLong(byte[] bb, long x, int index) {
             bb[index + 7] = (byte) (x >> 56);
@@ -73,15 +50,11 @@ public class ByteUtil {
             bb[index + 3] = (byte) (x >> 24);
             bb[index + 2] = (byte) (x >> 16);
             bb[index + 1] = (byte) (x >> 8);
-            bb[index + 0] = (byte) (x >> 0);
+            bb[index] = (byte) (x);
         }
 
         /**
          * 通过byte数组取到long
-         *
-         * @param bb
-         * @param index
-         * @return
          */
         public static long getLong(byte[] bb, int index) {
             return ((((long) bb[index + 7] & 0xff) << 56)
@@ -90,71 +63,56 @@ public class ByteUtil {
                     | (((long) bb[index + 4] & 0xff) << 32)
                     | (((long) bb[index + 3] & 0xff) << 24)
                     | (((long) bb[index + 2] & 0xff) << 16)
-                    | (((long) bb[index + 1] & 0xff) << 8) | (((long) bb[index + 0] & 0xff) << 0));
+                    | (((long) bb[index + 1] & 0xff) << 8) | (((long) bb[index] & 0xff)));
         }
 
         /**
          * 字符到字节转换
-         *
-         * @param ch
-         * @return
          */
         public static void putChar(byte[] bb, char ch, int index) {
             int temp = (int) ch;
             // byte[] b = new byte[2];
             for (int i = 0; i < 2; i ++ ) {
-                bb[index + i] = new Integer(temp & 0xff).byteValue(); // 将最高位保存在最低位
+                bb[index + i] = Integer.valueOf(temp & 0xff).byteValue(); // 将最高位保存在最低位
                 temp = temp >> 8; // 向右移8位
             }
         }
 
         /**
          * 字节到字符转换
-         *
-         * @param b
-         * @return
          */
         public static char getChar(byte[] b, int index) {
             int s = 0;
             if (b[index + 1] > 0)
                 s += b[index + 1];
             else
-                s += 256 + b[index + 0];
+                s += 256 + b[index];
             s *= 256;
-            if (b[index + 0] > 0)
+            if (b[index] > 0)
                 s += b[index + 1];
             else
-                s += 256 + b[index + 0];
-            char ch = (char) s;
-            return ch;
+                s += 256 + b[index];
+            return (char) s;
         }
 
         /**
          * float转换byte
-         *
-         * @param bb
-         * @param x
-         * @param index
          */
         public static void putFloat(byte[] bb, float x, int index) {
             // byte[] b = new byte[4];
             int l = Float.floatToIntBits(x);
             for (int i = 0; i < 4; i++) {
-                bb[index + i] = new Integer(l).byteValue();
+                bb[index + i] = Integer.valueOf(l).byteValue();
                 l = l >> 8;
             }
         }
 
         /**
          * 通过byte数组取得float
-         *
-         * @param b
-         * @param index
-         * @return
          */
         public static float getFloat(byte[] b, int index) {
             int l;
-            l = b[index + 0];
+            l = b[index];
             l &= 0xff;
             l |= ((long) b[index + 1] << 8);
             l &= 0xffff;
@@ -166,10 +124,6 @@ public class ByteUtil {
 
         /**
          * double转换byte
-         *
-         * @param bb
-         * @param x
-         * @param index
          */
         public static void putDouble(byte[] bb, double x, int index) {
             // byte[] b = new byte[8];
@@ -182,10 +136,6 @@ public class ByteUtil {
 
         /**
          * 通过byte数组取得float
-         *
-         * @param b
-         * @param index
-         * @return
          */
         public static double getDouble(byte[] b, int index) {
             long l;
@@ -196,13 +146,13 @@ public class ByteUtil {
             l |= ((long) b[2] << 16);
             l &= 0xffffff;
             l |= ((long) b[3] << 24);
-            l &= 0xffffffffl;
+            l &= 0xffffffffL;
             l |= ((long) b[4] << 32);
-            l &= 0xffffffffffl;
+            l &= 0xffffffffffL;
             l |= ((long) b[5] << 40);
-            l &= 0xffffffffffffl;
+            l &= 0xffffffffffffL;
             l |= ((long) b[6] << 48);
-            l &= 0xffffffffffffffl;
+            l &= 0xffffffffffffffL;
             l |= ((long) b[7] << 56);
             return Double.longBitsToDouble(l);
         }
